rename: 디자인 변경에 따른 디렉토리 구조 변경
- mobile, desktop 디자인이 상이하여 변경된 디렉토리 구조 반영
