+++ src/app/(home)/desktop/HomeDesktop.tsx
@@ -1,0 +1,14 @@
+import { Box } from "@mui/material"
+
+export default function HomeDesktop() {
+  return (
+    <Box
+      border={1}
+      sx={{
+        display: { xs: "none", sm: "flex" },
+      }}
+    >
+      HomeDesktop
+    </Box>
+  )
+}
