+++ src/app/(home)/mobile/HomeMobile.tsx
@@ -1,0 +1,14 @@
+import { Box } from "@mui/material"
+
+export default function HomeMobile() {
+  return (
+    <Box
+      border={1}
+      sx={{
+        display: { xs: "flex", sm: "none" },
+      }}
+    >
+      HomeMobile
+    </Box>
+  )
+}
