+++ src/app/components/common/BlurImage.tsx
